pyproject.toml
[project]
name = "kr-opendata-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=3.0.2",
    "pymysql>=1.1.2",
    "python-dotenv>=1.0.0",
    "requests>=2.33.1",
]
